--- atomicwriter.egg-info/PKG-INFO
@@ -1,16 +1,0 @@
-Metadata-Version: 2.4
-Name: atomicwriter
-Version: 0.1.0
-Summary: Atomic file writer for Pandas, Polars, Numpy, etc.
-Author: Your Name
-License: Apache-2.0
-Requires-Python: >=3.7
-License-File: LICENSE
-Requires-Dist: pandas
-Requires-Dist: pyarrow
-Dynamic: author
-Dynamic: license
-Dynamic: license-file
-Dynamic: requires-dist
-Dynamic: requires-python
-Dynamic: summary
