--- atomicwriter.egg-info/requires.txt
@@ -1,2 +1,0 @@
-pandas
-pyarrow
