+++ pyproject.toml
@@ -1,3 +1,37 @@
 [build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta" 
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "atomicwriter"
+version = "0.1.0"
+authors = [
+    { name="Your Name", email="you@example.com" },
+]
+description = "Atomic file writer for Pandas, Polars, Numpy, etc."
+readme = "README.md"
+license = { text = "Apache-2.0" }
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent",
+]
+
+# setup.py와 requirements.txt의 핵심 의존성을 여기에 통합
+dependencies = [
+    "pandas",
+    "pyarrow",
+    "polars",
+    # "fsspec", # 나중에 원격 지원 추가 시 주석 해제
+]
+
+# (선택 사항이지만 강력 추천) 개발용 의존성 분리
+[project.optional-dependencies]
+dev = [
+    "pytest",
+]
+
+# `find_packages()` 역할을 하는 부분 (src 레이아웃 사용 시)
+[tool.setuptools]
+packages = { find = { where = ["src"] } }
