--- requirements.txt
@@ -1,4 +1,0 @@
-pandas 
-pyarrow 
-pytest
-polars
