+++ src/atomicwriter/__init__.py
